Move promotion-specific code from Spree::Adjustment to legacy_promotions
We don't need #recalculate for anything in core, and we don´t need the
association to promotion codes.
